+++ include/liberation_x3g.h
@@ -8,12 +8,27 @@
 #define X3G_MAX_OBJECTS    64
 #define X3G_MAX_VERTICES  256
 #define X3G_MAX_POLYGONS  256
+#define X3G_MAX_POLY_VERTS 8
 
 typedef struct {
     int16_t x, y, z;
     int16_t group;
     int16_t reserved[4];
 } X3gVertex;
+
+typedef struct {
+    uint16_t type;
+    uint16_t record_size;
+    int16_t  w2, w3, w4, w5;
+    int16_t  w6, w7;
+    int16_t  normal_x, normal_y;
+    uint16_t flags;
+    uint16_t color;
+    uint16_t uv_left, uv_top, uv_right, uv_bottom;
+    uint16_t w16;
+    uint8_t  vertex_count;
+    uint8_t  vertex_indices[X3G_MAX_POLY_VERTS];
+} X3gPolygon;
 
 typedef struct {
     const uint8_t *data;
@@ -24,6 +39,8 @@ typedef struct {
     X3gVertex      *vertices;
     unsigned        vertex_count;
     X3gPolygonList  polygons;
+    X3gPolygon     *parsed_polys;
+    unsigned        polygon_count;
     const uint8_t  *extra;
     unsigned        extra_size;
 } X3gObject;